requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
openai>=1.40.0,<2.0.0
tiktoken>=0.7.0,<1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
tqdm==4.66.2
pypdf==4.2.0
langchain-core>=0.3.54
langchain-openai>=0.2.0
langchain-community>=0.0.21
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.1
chromadb>=0.4.24
langgraph>=0.0.20
schedule==1.2.1
lxml==5.1.0
